requirements.txt
playwright==1.41.0
openai>=1.0.0
python-dotenv>=1.0.0
